# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "View all users",
                "operationId": "069866150be4b975d248c35627c94fdb",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "required|string"
                                        },
                                        "email": {
                                            "type": "required|string|email"
                                        },
                                        "password": {
                                            "type": "required|string"
                                        },
                                        "created_at": {
                                            "type": "required|string|dateTime"
                                        },
                                        "updated_at": {
                                            "type": "required|string|dateTime"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": [
                                            {
                                                "id": 1,
                                                "name": "Andreew",
                                                "email": "andreew.k.januario@outlook.com",
                                                "email_verified_at": null,
                                                "created_at": "2024-01-11T15:19:20.000000Z",
                                                "updated_at": "2024-01-11T15:19:20.000000Z"
                                            },
                                            {
                                                "id": 4,
                                                "name": "Bill",
                                                "email": "bill.k.januario@outlook.com",
                                                "email_verified_at": null,
                                                "created_at": "2024-01-11T15:19:20.000000Z",
                                                "updated_at": "2024-01-11T15:19:20.000000Z"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "operationId": "14ffc1a2beec11e409c05b16801dd1c1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "required|string"
                                    },
                                    "email": {
                                        "type": "required|string|email"
                                    },
                                    "password": {
                                        "type": "required|string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "fulano",
                                    "email": "fulano@gmail.com",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": {
                                            "message": "Usuário cadastrado com sucesso!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}